fix(wv/windows): default-deny camera and microphone via keld-guard
KEL-59 closed the media-capture hole on macOS but left Windows on the
platform default, so Keld's headline property — default-deny, host-enforced —
only held on one OS. A capability model that is platform-conditional is not a
capability model.

Windows was not vulnerable the way macOS was. Verified against Microsoft's
reference for CoreWebView2PermissionState: `Default` (0x0) means "the default
browser behavior is used, which normally prompts users for decision", so
WebView2 asks rather than auto-granting the way wry's WKWebView delegate did.
But default-ask is still wrong here: a prompt lets a renderer obtain capture
the manifest never granted, and arch 03 §1 makes the manifest the authority.
Deny is therefore deliberate, not a copy of the macOS fix.

Rather than duplicate the mapping into a second backend, the three wry helpers
move to media.rs behind a cfg for the platforms that link wry. Duplicating a
security decision is how the two copies drift; root AGENTS.md calls duplicated
helpers slop. Their tests move with them, so the wry kind mapping and the
Deny-not-Default assertions now run on Windows instead of macOS only.

The source-guard test is generalised from macos_backend_installs_guarded_handler
to backends_install_guarded_handler and loops over both live backends, because
dropping the call is silent: macOS falls back to Grant, Windows to a prompt,
and no other test notices.

Gates: fmt clean; clippy --workspace --all-targets -D warnings clean;
cargo test --workspace 223 passed 0 failed.

Refs: KEL-59, KEL-27
